# docs/swagger.yaml
basePath: /api
definitions:
  internal_handler.AttachmentResponse:
    properties:
      contentType:
        type: string
      entityId:
        type: string
      entityType:
        type: string
      expiresAt:
        type: string
      fileName:
        type: string
      fileSize:
        type: integer
      fileUrl:
        type: string
      id:
        type: string
      status:
        type: string
      uploadedAt:
        type: string
      uploadedBy:
        type: string
    type: object
  internal_handler.PresignedURLRequest:
    properties:
      contentType:
        type: string
      entityType:
        type: string
      fileName:
        type: string
      fileSize:
        type: integer
      workspaceId:
        type: string
    required:
    - contentType
    - entityType
    - fileName
    - fileSize
    - workspaceId
    type: object
  internal_handler.PresignedURLResponse:
    properties:
      attachmentId:
        type: string
      expiresIn:
        description: seconds
        type: integer
      fileKey:
        type: string
      uploadUrl:
        type: string
    type: object
  internal_handler.SaveAttachmentMetadataRequest:
    properties:
      contentType:
        type: string
      entityType:
        type: string
      fileKey:
        type: string
      fileName:
        type: string
      fileSize:
        type: integer
    required:
    - contentType
    - entityType
    - fileKey
    - fileName
    - fileSize
    type: object
...
  project-board-api_internal_response.ErrorResponse:
    properties:
      error: {}
      requestId:
        type: string
    type: object
  project-board-api_internal_response.SuccessResponse:
    properties:
      data: {}
      requestId:
        type: string
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    프로젝트 보드 관리 시스템 API 서버입니다.
    Board, Project, Comment, Participant 관리 기능을 제공합니다.
    ALB path-based routing with /api/boards prefix
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Project Board Management API
  version: "1.0"
paths:
  /attachments:
    post:
      consumes:
      - application/json
      description: |-
        Saves attachment metadata to the database after successful S3 upload
        Creates a temporary attachment record with 1-hour expiration
        The attachment will be linked to an entity (board/comment/project) when that entity is created
        Supported entity types: BOARD, COMMENT, PROJECT
      parameters:
      - description: Attachment metadata
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SaveAttachmentMetadataRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Attachment metadata saved successfully
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.AttachmentResponse'
              type: object
        "400":
          description: Invalid request or file validation failed
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "401":
          description: Unauthorized - user not authenticated
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: Failed to save attachment metadata
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Save attachment metadata
      tags:
      - attachments
  /attachments/{attachmentId}:
    delete:
      consumes:
      - application/json
      description: |-
        Deletes an attachment from both S3 and database
        Only the user who uploaded the attachment can delete it
        Performs soft delete on the database record
      parameters:
      - description: Attachment ID
        in: path
        name: attachmentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Attachment deleted successfully
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Invalid attachment ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "401":
          description: Unauthorized - user not authenticated
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "403":
          description: Forbidden - user does not have permission to delete this attachment
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Attachment not found
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: Failed to delete attachment
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Delete attachment
      tags:
      - attachments
  /attachments/presigned-url:
    post:
      consumes:
      - application/json
      description: |-
        Generates a presigned URL for uploading a file directly to S3
        Creates a temporary attachment record and returns its ID along with the presigned URL
        Validates file metadata (size, type, name) before generating URL
        Supported entity types: BOARD, COMMENT, PROJECT
        Supported file types: images (jpg, jpeg, png, gif, webp) and documents (pdf, txt, doc, docx, xls, xlsx, ppt, pptx)
        Maximum file size: 20MB
        URL expires in 5 minutes (300 seconds)
      parameters:
      - description: Presigned URL request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.PresignedURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Presigned URL generated successfully
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.PresignedURLResponse'
              type: object
        "400":
          description: Invalid request or file validation failed
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "401":
          description: Unauthorized - user not authenticated
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: Failed to generate presigned URL
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Generate presigned URL for file upload
      tags:
      - attachments
...
  /boards/{boardId}/attachments:
    get:
      consumes:
      - application/json
      description: |-
        Retrieves all attachments associated with a specific board
        Returns only confirmed attachments linked to the board
      parameters:
      - description: Board ID
        in: path
        name: boardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Attachments retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handler.AttachmentResponse'
                  type: array
              type: object
        "400":
          description: Invalid board ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: Failed to retrieve attachments
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Get board attachments
      tags:
      - attachments
...
  /comments/{commentId}/attachments:
    get:
      consumes:
      - application/json
      description: |-
        Retrieves all attachments associated with a specific comment
        Returns only confirmed attachments linked to the comment
      parameters:
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Attachments retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handler.AttachmentResponse'
                  type: array
              type: object
        "400":
          description: Invalid comment ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: Failed to retrieve attachments
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Get comment attachments
      tags:
      - attachments
...
  /projects/{projectId}/attachments:
    get:
      consumes:
      - application/json
      description: |-
        Retrieves all attachments associated with a specific project
        Returns only confirmed attachments linked to the project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Attachments retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handler.AttachmentResponse'
                  type: array
              type: object
        "400":
          description: Invalid project ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: Failed to retrieve attachments
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Get project attachments
      tags:
      - attachments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT 토큰을 입력하세요. 형식: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"